Update and rename trajectory.py to Trajectory.py
Revamped. No longer uses in-house RK4 of RKF45. Uses Scipy's odeint (which makes it run much quicker). Plots the trajectory overlayed on the zero-velocity curve corresponding to the Jacobi constant chosen. It also plots the Jacobi constant over the time of the trajectory to see if it is actually constant. Spoiler: Pretty much, yeah it is. Unless you consider deviation 1e-6 not constant.